refactor: ItemRepository 디렉토리 이동
