+++ src/main/java/saphy/saphy/item/domain/repository/ItemRepository.java
@@ -1,4 +1,4 @@
-package saphy.saphy.item.repository;
+package saphy.saphy.item.domain.repository;
 
 import java.util.List;
